Modified the code such that all generated species + reactions are considered

+++ BioChemSys1.1/src/ro/ubb/biochem/reaction/components/RuleRepository.java
@@ -16,7 +16,7 @@ public interface RuleRepository {
 	
 	public List<Rule> getRulesForLhs(List<Specie> species);
 
-	public Double getNumberOfRules();
+	public int getNumberOfRules();
 	
 	public List<Rule> getRandomSetOfRules(Integer requiredSetSize);
